+++ packages/amplify-e2e-core/src/configure/index.ts
@@ -32,7 +32,6 @@ export const amplifyRegions = [
   'ap-southeast-2',
   'ap-south-1',
   'ca-central-1',
-  'me-south-1',
   'sa-east-1',
 ];