+++ packages/amplify-data-construct/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.17.3](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.17.2...@aws-amplify/data-construct@1.17.3) (2026-05-07)
+
+### Bug Fixes
+
+- add missing transitive bundled dependencies to construct packages ([#3473](https://github.com/aws-amplify/amplify-category-api/issues/3473)) ([abe1ac7](https://github.com/aws-amplify/amplify-category-api/commit/abe1ac73f9cd8e124292f70cae39eef15aa10995)), closes [aws-amplify/amplify-backend#3158](https://github.com/aws-amplify/amplify-backend/issues/3158) [aws-amplify/amplify-backend#3158](https://github.com/aws-amplify/amplify-backend/issues/3158)
+- dep updates ([#3441](https://github.com/aws-amplify/amplify-category-api/issues/3441)) ([bddb2c8](https://github.com/aws-amplify/amplify-category-api/commit/bddb2c84dfb1288d45f91f852f8b48207e10aecd)), closes [#295](https://github.com/aws-amplify/amplify-category-api/issues/295) [#294](https://github.com/aws-amplify/amplify-category-api/issues/294) [#302](https://github.com/aws-amplify/amplify-category-api/issues/302) [#307-310](https://github.com/aws-amplify/amplify-category-api/issues/307-310) [#236](https://github.com/aws-amplify/amplify-category-api/issues/236) [#273-275](https://github.com/aws-amplify/amplify-category-api/issues/273-275) [#282-285](https://github.com/aws-amplify/amplify-category-api/issues/282-285) [#290](https://github.com/aws-amplify/amplify-category-api/issues/290) [#292](https://github.com/aws-amplify/amplify-category-api/issues/292)
+
+## [1.17.2](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.17.1...@aws-amplify/data-construct@1.17.2) (2026-04-13)
+
+**Note:** Version bump only for package @aws-amplify/data-construct
+
 ## [1.17.1](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/data-construct@1.17.0...@aws-amplify/data-construct@1.17.1) (2026-03-24)
 
 ### Bug Fixes